[core][Android] Removes `compose` integration if not needed
