+++ packages/expo-modules-core/android/build.gradle
@@ -2,25 +2,30 @@ import expo.modules.plugin.gradle.ExpoModuleExtension
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 buildscript {
+  // List of features that are required by linked modules
+  def coreFeatures = project.findProperty("coreFeatures") ?: []
+  ext.shouldIncludeCompose = coreFeatures.contains("compose")
+
   repositories {
     mavenCentral()
   }
 
   dependencies {
-    classpath("org.jetbrains.kotlin.plugin.compose:org.jetbrains.kotlin.plugin.compose.gradle.plugin:${kotlinVersion}")
+    if (shouldIncludeCompose) {
+      classpath("org.jetbrains.kotlin.plugin.compose:org.jetbrains.kotlin.plugin.compose.gradle.plugin:${kotlinVersion}")
+    }
   }
 }
 
 apply plugin: 'com.android.library'
 apply plugin: 'expo-module-gradle-plugin'
-apply plugin: 'org.jetbrains.kotlin.plugin.compose'
 
+if (shouldIncludeCompose) {
+  apply plugin: 'org.jetbrains.kotlin.plugin.compose'
+}
 
 group = 'host.exp.exponent'
 version = '2.3.7'
-
-// List of features that are required by linked modules
-def coreFeatures = project.findProperty("coreFeatures") ?: []
 
 def isExpoModulesCoreTests = {
   Gradle gradle = getGradle()
@@ -98,7 +103,7 @@ android {
   buildFeatures {
     buildConfig true
     prefab true
-    compose true
+    compose shouldIncludeCompose
   }
 
   packagingOptions {
@@ -155,6 +160,12 @@ android {
           // TODO(kudo,20241112): Remove this when we drop react-native 0.76 support
           srcDirs += 'src/rn74'
         }
+
+        if (shouldIncludeCompose) {
+          srcDirs += 'src/compose'
+        } else {
+          srcDirs += 'src/withoutCompose'
+        }
       }
     }
   }
@@ -183,7 +194,9 @@ dependencies {
   api "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3"
   api "androidx.core:core-ktx:1.13.1"
 
-  implementation 'androidx.compose.foundation:foundation-android:1.7.6'
+  if (shouldIncludeCompose) {
+    implementation 'androidx.compose.foundation:foundation-android:1.7.6'
+  }
 
   implementation("androidx.tracing:tracing-ktx:1.2.0")
