+++ packages/expo-modules-core/android/src/compose/expo/modules/kotlin/views/AutoSizingComposable.kt
