+++ packages/expo-modules-core/android/src/withoutCompose/expo/modules/kotlin/views/ModuleDefinitionBuilderComposeExtension.kt
@@ -1,0 +1,10 @@
+@file:Suppress("FunctionName")
+
+package expo.modules.kotlin.views
+
+import expo.modules.kotlin.modules.InternalModuleDefinitionBuilder
+import expo.modules.kotlin.modules.Module
+
+open class ModuleDefinitionBuilderWithCompose(
+  module: Module? = null
+) : InternalModuleDefinitionBuilder(module)
